Move generic visitor class to tools so it can be shared between modules
